+++ R/standard_error.R
@@ -1266,6 +1266,22 @@ standard_error.rma <- function(model, ...) {
 
 
 
+#' @export
+standard_error.lavaan <- function(model, ...) {
+  out <- .extract_parameters_lavaan(model, ...)
+  out[out$Operator != "~1", c("To", "Operator", "From", "SE")]
+}
+
+
+
+#' @export
+standard_error.blavaan <- function(model, ci = .95, ...) {
+  out <- .extract_parameters_blavaan(model, ...)
+  out[out$Operator != "~1", c("To", "Operator", "From", "SE")]
+}
+
+
+